cremailleurs-api/requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==36.0.2
fake-useragent==0.1.11
Flask==2.0.3
Flask-Cors==3.0.10
h11==0.13.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
mysql-connector-python==8.0.28
outcome==1.1.0
protobuf==3.19.4
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
requests==2.27.1
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
stem==1.8.0
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
webdriver-manager==3.5.4
Werkzeug==2.0.3
wsproto==1.1.0
